+++ src/collections/models.py
@@ -1,6 +1,7 @@
-from ..utils.models import Model
-from ..utils.errors import ModelError
 from random import randint
+
+from src.utils.base.errors import ModelError
+from src.utils.base.models import Model
 
 
 class CollectionResultSet(Model):
