+++ src/members/views.py
@@ -1,10 +1,11 @@
-from flask import jsonify, json, request
-from flask.json import loads
-from flask.views import MethodView
-from ..utils.errors import *
-from .models import *
-from flask import current_app
 import traceback
+
+from flask import current_app
+from flask import json, request
+from flask.views import MethodView
+
+from src.utils.base.errors import *
+from .models import *
 
 
 class MemberView(MethodView):
