+++ src/utils/ids/url_encoder.py
@@ -1,5 +1,7 @@
 import base64
-from .struct import Struct
+
+from src.utils.base.struct import Struct
+
 encoder = Struct(
     encode = lambda s: base64.b64encode(str.encode(s)).decode(),
     decode = lambda s: base64.b64decode(s).decode()
