+++ src/utils/rdf/queries.py
