+++ test-env.py
@@ -1,9 +1,4 @@
-import base64
-from rdflib import Graph, URIRef, Namespace, Literal
-from rdflib.namespace import RDF, DCTERMS
-from src.data.ldp_db import *
-from src.collections.models import *
-from src.members.models import *
+from src.utils.data.ldp_db import *
 from test.mock import RandomGenerator
 
 mocks = RandomGenerator()
